Add C++ driver for the Grove RGB LCD
A first version of a C++ driver implementation for the Grove RGB LCD.
Due to issues with the LCD device, just for the RGB part.
As the C++ implementation binds on the exactly same device as the C one
(the drivers needs to be compareable for the thesis), has the same
functional scope and uses the same FIDL interface, only one
implementation is allowed to build at the same time.
Even if there is no compile error it would lead to runtime effects.

The implementation tries to use C++ as idiomatic as possible in the
given context, but may needs improvement. As there is no documentation
for C++ drivers in Zircon and the C bindings are compatible with C++,
there are several ways and styles out there to use, but no best practise.

The C++ driver need to perform the first binding steps in C due to current?
limitations in Zircon, before the handing over to C++ is possible.