Add PDev variants for the C- and Cpp-Driver
The HiKey960 is an ARM platform, there is no standardized, queryable
bus containing e.g. the I2C host controller. That's the reason for
using the "platform bus", a pseudo bus.
As you can not query this bus for the attached devices, as the I2C
controller, it is needed to know about them and share this knowledge
with the kernel. For Zircon, the way to share are "board files" which
define the devices, their addresses and so on in source code.
Zircon allows to group more than one I2C slave description in a board
file and define a single DID (device id) for them. In driver
development, this means you can bind more than one I2C slave device on
a single driver. That's really cool for devices made from two or more
slave addresses, exactly like the Grove LCD RGB backlight display.
The added drivers show way for C and Cpp to implement such a platform
device I2C driver.