Add LCD driver implementation
After the technical issues are solved (the LCD controller on the Grove
LCD was not ok with the levelshifting), the driver implementation for
the LCD started. It adds basic "read" and "write" implementations, but
we expect the use via FIDL.
To avoid padding issues, the slave controller's target register for
write operations (0x40) is attached to the written strings as a prefix
(@).
The FIDL definition is relocated to a generic folder for both interface
definitions, the one for RGB and the one for LCD. It is not tested if
this approach is applicable for the PDev drivers.
A FIDL interface for the LCD is defined. It is rather a proof of concept
than a complete interface for all functionality, the LCD provide.
The definitions are already added to the "grove-c-driver", but not
implemented right now.