Add LCD implementation for C-PDev variant
Adds the implementation for the LCD in the C driver's PDev variant.
As it is only possible to implement one FIDL definition at a time, this
driver includes a new, combined variante for both actual Grove devices
especially used for the PDev drivers.
The "read()" and "write()" system calls are one issue for a Grove driver
in the PDev variant. Much more as in the normal ones, it is not that
clear what especially "write()" should do exactly. It could be solved by
massive string parsing and defined commands, a rather bad option.
This is exactly the situation FIDL was created for. FIDL enables us to
create meaningful, clean APIs to interact with the Grove-LCD RGB
backlight, so there is no valid reason for implement bad code just to
provide "read()" and "write()". It is a bit the same for the normal
variants, but not that much as in the PDev ones. So we implemented
these system calls in the normal ones, but rather as a show case than as
a valid way to interact.