Add basic user apps for testing Grove FIDL calls
These are just very basic user apps for short tests on the Grove-LCD RGB
driver's FIDL implementations. As the normal and PDev variants use
different FIDL definitions and result in different device file paths, we
duplicated the test implementation. Build more sophisticated user apps
is a ToDo for a later time in this thesis (if there is enough time to do
so).