Add first implementation for Grove LCD Cpp Driver
Added the most of the implementation for the LCD part of the Cpp driver.
Due to linker and make file issues, there is some further work needed.
For the same reason it is not tested at the moment and the
implementations for the "read()" and "write()" system calls are still
missing.