Add LCD implementation for Grove C++ driver
Adds an implementation for the Grove C++ LCD driver part, using
fbl::String for string handling. Put fbl::String in the RGB part as much
as possible to get rid of the C Strings.
Fixes some issues in the rules.mk file.
The C++ driver itself is functionally the same as the C variant. For
"read" and "write", the driver inherit from Ddk::Readable and
Ddk::Writeable. Same for the FIDL Message system.
The FIDL definitions are at the moment just available for C (C
bindings). That means it is a bit ugly to use/implement this definition
in a proper C++ driver and matching style. It is done by reinterpreting
the context pointer as "this" object.