Add LCD implementation for the Grove C++ PDev driver
The C++ PDev driver is functionally exactly the same as the other PDev
variants. In contrast to the normal ones, the PDev variants do not
implement the "read()" and "write()" system calls. The reasons are
already listed in the commit message for the C PDev driver.

The C++ PDev driver's implementation is in fact the same as the normal C++
ones, but in contrast to the other one, it is just one driver managing
both partial I2C devices. For this, the driver take advantage of running
on a "platform device" and use this special APIs. The concrete PDev itself
is allocated via inheritance and based on this, the I2C channels are
allocated via the PDev API.

As there are some difficulties with initialization threads at C++
drivers are the ones in the C++ drivers running just as normal function
calls.